pyproject.toml
[tool.poetry.scripts]
app = "app.__main__:main"
start = "app.__main__:main"

[project]
name = "university-finance-api"
version = "0.1.0"
description = ""
authors = [
    {name = "beerappa88",email = "beerappametre01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9, <3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    # packages = []
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "bcrypt (==4.0.1)",
    "pyotp (>=2.9.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

